(RD)
changed added a empty cart activity for when the card is empty
added a if state in menu that checks if the cart is empty or full
if the cart is empty it goes to EmptyCart activity if its full it'll go to ShoppingCart.
also change the fab icons for when the user wants to go into the cart and when their in a full cart to complete their order.